LayoutTests/http/tests/cache/disk-cache/disk-cache-redirect-to-data-expected.txt

Test that HTTP redirect to a data URL is blocked for subresources.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS self.redirectedScriptExecuted is false
PASS successfullyParsed is true

TEST COMPLETE
